chore : img태그 수정, 랜딩 로고 짤림 수정
