+++ next-netfilx/src/components/MainPage/MainButtons.tsx
@@ -1,17 +1,27 @@
 import styled from "styled-components";
+import Image from "next/image";
 
 export default function MainButtons() {
   return (
-    <>
-      <ButtonContainer>
-        <img src="/myList.svg" alt="" />
-        <PlayBox>
-          <img src="/playButton.svg" alt="" />
-          <img src="/play.svg" alt="" />
-        </PlayBox>
-        <img src="/infoIcon.svg" alt="" />
-      </ButtonContainer>
-    </>
+    <ButtonContainer>
+      <Image src="/myList.svg" width={41} height={45} alt="My List" />
+      <PlayBox>
+        <Image
+          src="/playButton.svg"
+          width={110}
+          height={45}
+          alt="Play Button Background"
+        />
+        <Image
+          src="/play.svg"
+          width={72}
+          height={30}
+          alt="Play Icon"
+          className="play-icon"
+        />
+      </PlayBox>
+      <Image src="/infoIcon.svg" width={24} height={45} alt="Info" />
+    </ButtonContainer>
   );
 }
 
@@ -21,7 +31,6 @@ const ButtonContainer = styled.div`
   justify-content: space-around;
   width: 80%;
   margin-top: 11px;
-  position: static;
 `;
 
 const PlayBox = styled.div`
@@ -29,10 +38,10 @@ const PlayBox = styled.div`
   height: 45px;
   position: relative;
 
-  img:nth-child(2) {
+  .play-icon {
     position: absolute;
     top: 50%;
     left: 50%;
-    transform: translateX(-50%) translateY(-50%);
+    transform: translate(-50%, -50%);
   }
 `;
